register user

+++ src/main/java/com/cojo/passwordmanager/EncryptedDataRepository.java
@@ -2,6 +2,6 @@ package com.cojo.passwordmanager;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EncryptedDataRepository extends JpaRepository<EncryptedData, Integer> {
+public interface EncryptedDataRepository extends JpaRepository<EncryptedData, Long> {
     
 }
